+++ CHANGES.md
@@ -1,3 +1,5 @@
 - Bugfix on Outdated version display of changes.
 - Bugfix in Virtual Terminal in cents values
 - Enable HTML as Plugin response content
+- Bugfix of Cookie issue
+- Enable Partner Settings